Add (empty) test suite
