+++ plutus-benchmark/values/test/Spec.hs
@@ -1,0 +1,8 @@
+module Main where
+
+import PlutusBenchmark.Values.FlattenedValue qualified as FlattenedValue
+import PlutusBenchmark.Values.MMValue qualified as MMValue
+import PlutusBenchmark.Values.NestedValue qualified as NestedValue
+
+main :: IO ()
+main = undefined
